refactor: remove unneeded decorations and windows impl
