+++ src/protocol/index.ts
@@ -2,7 +2,6 @@ export * from './capabilities'
 export * from './command'
 export * from './configuration'
 export * from './contribution'
-export * from './decoration'
 export * from './definition'
 export * from './hover'
 export * from './implementation'
